Touchups for Mac and twilight (#262)
* update socket.io

* add no-run and enable debugging

* Revert "update socket.io"

This reverts commit f26dad9e6285aa2891eab5160fe14546157a4fa4.

* support no_delete_runs

* disable js debugging

* don't use is, add debugging to Trace

* fix bug with dot using tabs

+++ middleware/qira_program.py
@@ -285,10 +285,11 @@ class Program:
     else:
       raise Exception("unknown binary type")
 
-  def clear(self):
+  def clear(self, delete_old_runs=True):
     # probably always good to do except in development of middleware
-    print("*** deleting old runs")
-    self.delete_old_runs()
+    if delete_old_runs:
+      print("*** deleting old runs")
+      self.delete_old_runs()
 
     # getting asm from qemu
     self.create_asm_file()
